Add AuthorityQuery, InquestQuery types and Joi schemas
