--- src/constants.ts
@@ -1,7 +1,0 @@
-export const PAGINATION = 12;
-
-export enum Sort {
-  Relevant = 'relevant',
-  New = 'new',
-  Alphabetical = 'a-z',
-}
